# openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "description": "One envelope for every error response: `{ \"code\", \"message\" }`.",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "ConfigAppliedDto": {
        "description": "Result of a direct (non-sensitive) apply.",
        "properties": {
          "changed_keys": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "generation": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "generation",
          "changed_keys"
        ],
        "type": "object"
      },
      "ConfigEntryDto": {
        "description": "One committed config entry.",
        "properties": {
          "key": {
            "type": "string"
          },
          "value": {}
        },
        "required": [
          "key",
          "value"
        ],
        "type": "object"
      },
      "ConfigProposalDto": {
        "properties": {
          "base_generation": {
            "format": "int64",
            "type": "integer"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "patch": {},
          "reason": {
            "type": "string"
          },
          "resulting_generation": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "base_generation",
          "patch",
          "reason",
          "expires_at"
        ],
        "type": "object"
      },
      "ConfigSnapshotDto": {
        "description": "The complete committed snapshot plus its generation watermark.",
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/ConfigEntryDto"
            },
            "type": "array"
          },
          "generation": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "generation",
          "entries"
        ],
        "type": "object"
      },
      "CreateConfigProposalRequest": {
        "description": "Two-phase proposal creation for sensitive keys (D24). Exactly one of\n`patch` / `revert_of_generation` must be present: the latter pre-fills an\nemergency-revert patch from that generation's `config_changes` history\n(`old` values; pause keys stripped) under the SAME dual control.",
        "properties": {
          "idempotency_key": {
            "type": "string"
          },
          "patch": {},
          "reason": {
            "type": "string"
          },
          "revert_of_generation": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "reason",
          "idempotency_key"
        ],
        "type": "object"
      },
...
      "SetConfigRequest": {
        "description": "Typed patch over the prospective snapshot (D24 write path). Non-sensitive\nkeys apply directly; sensitive keys are rejected toward the proposal flow.",
        "properties": {
          "expected_base_generation": {
            "description": "Optimistic base; a moved generation is a typed conflict.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "idempotency_key": {
            "type": "string"
          },
          "patch": {
            "description": "Map of key → new value."
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "patch",
          "reason",
          "idempotency_key"
        ],
        "type": "object"
      },
      "SettleConfigProposalRequest": {
        "description": "Confirm/reject body: the acting principal comes from RBAC, the reason is\nrecorded in the audit fact.",
        "properties": {
          "reason": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "",
    "license": {
      "name": ""
    },
    "title": "adapters",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/config": {
      "get": {
        "operationId": "get_config",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigSnapshotDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "ops"
        ]
      },
      "post": {
        "operationId": "set_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigAppliedDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "ops"
        ]
      }
    },
    "/admin/config/proposals": {
      "post": {
        "operationId": "create_proposal",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConfigProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigProposalDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "ops"
        ]
      }
    },
    "/admin/config/proposals/{id}/confirm": {
      "post": {
        "operationId": "confirm_proposal",
        "parameters": [
          {
            "description": "Proposal id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettleConfigProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigProposalDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "ops"
        ]
      }
    },
    "/admin/config/proposals/{id}/reject": {
      "post": {
        "operationId": "reject_proposal",
        "parameters": [
          {
            "description": "Proposal id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettleConfigProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigProposalDto"
                }
              }
            },
            "description": ""
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "ops"
        ]
      }
    },
...
  },
...
}